requirements.txt
gputil==1.4.0
matplotlib==3.10.9
mlflow==3.7.0
mlflow_skinny==3.7.0
mlflow_tracing==3.7.0
numpy==2.4.4
opencv_python_headless==4.12.0.88
pandas==3.0.2
Pillow==12.2.0
psutil==7.1.3
PyYAML==6.0.3
scipy==1.17.1
seaborn==0.13.2
Shapely==2.1.2
skimage==0.0
tifffile==2025.12.12
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
